refactor(daily): split daily table form and stabilize router smoke
